chore(lint): split over-budget files / functions / closures
SwiftLint thresholds — file_length, type_body_length, function_body_length,
closure_body_length, large_tuple, contains_over_first_not_nil — all pass
again at the source rather than via baseline edits.

Splits and extractions:
- SwiftDataToGRDBMigrator+CoreFinancialGraph split: earmarks /
  earmark_budget_items / investment_values move to a new sibling
  SwiftDataToGRDBMigrator+Earmarks.
- SwiftDataToGRDBMigratorTests split into three siblings — the
  per-record-type core-graph tests move to
  SwiftDataToGRDBMigratorCoreGraphTests, the cross-FK end-to-end test
  moves to SwiftDataToGRDBMigratorCrossFKTests; the original keeps the
  CSV-import / import-rule tests.
- CoreFinancialGraphSyncRoundTripTests' transaction + leg tests move
  to SyncRoundTripTransactionTests with a shared seedLegParents helper.
- CoreFinancialGraphRollbackTests' category-delete fixture extracted
  into nonisolated helpers (CategoryDeleteFixtureIds struct,
  insertCategoryDeleteParents, insertCategoryDeleteTransaction,
  seedCategoryDeleteFixture orchestrator) to drop the long test body
  and its inner closure under threshold; replaced
  `first(where:) != nil` with `contains(where:)`.
- InstrumentRegistryContractTests.makeSubject() returns a Subject
  struct (repo / hooks / database) instead of a 3-tuple.
- TestBackend.seed(transactions:) closure extracted ensureLegParents
  and a TransactionSeedState struct so the write-closure body fits
  the 30-line ceiling.

just build-mac green. just test-mac green: 1727 tests across 263
suites pass. just format-check clean.
