+++ MoolahTests/Backends/GRDB/CoreFinancialGraphRollbackTests.swift
@@ -83,7 +83,7 @@ struct CoreFinancialGraphRollbackTests {
     }
     let surviving = try #require(accounts.first { $0.id == priorId })
     #expect(surviving.name == "prior")
-    #expect(accounts.first { $0.id == failingId } == nil)
+    #expect(!accounts.contains(where: { $0.id == failingId }))
   }
 
   // MARK: - GRDBTransactionRepository.create(_:)
@@ -141,72 +141,34 @@ struct CoreFinancialGraphRollbackTests {
 
   // MARK: - GRDBCategoryRepository.delete(id:withReplacement:)
 
-  @Test
-  func categoryDeleteRollsBackOnFailure() async throws {
-    let database = try ProfileDatabase.openInMemory()
-    let categoryRepo = GRDBCategoryRepository(database: database)
+  /// Identifiers seeded ahead of `categoryDeleteRollsBackOnFailure` and
+  /// re-checked after the failed delete to assert nothing torn through.
+  private struct CategoryDeleteFixtureIds {
     let accountId = UUID()
     let categoryId = UUID()
     let earmarkId = UUID()
     let txnId = UUID()
     let legId = UUID()
     let budgetItemId = UUID()
-    let category = Moolah.Category(id: categoryId, name: "Food")
+  }
+
+  /// Seeds the parent rows + leg + budget item that reference the
+  /// category, then installs the BEFORE-DELETE trigger on
+  /// `earmark_budget_item` so the production delete path's
+  /// budget-item DELETE step throws.
+  ///
+  /// The category's delete path UPDATEs the leg (category_id → NULL),
+  /// DELETEs the budget item, then DELETEs the category. A failing
+  /// trigger on the budget-item DELETE must unwind the leg UPDATE along
+  /// with the row removal.
+  private func seedCategoryDeleteFixture(
+    in database: any DatabaseWriter,
+    ids: CategoryDeleteFixtureIds
+  ) async throws {
+    let category = Moolah.Category(id: ids.categoryId, name: "Food")
     try await database.write { database in
-      // Seed the parent rows and a leg+budget item that reference the
-      // category. The category's delete path UPDATEs the leg
-      // (category_id → NULL), DELETEs the budget item, then DELETEs the
-      // category. A failing trigger on the budget-item DELETE must
-      // unwind the leg UPDATE along with the row removal.
-      try AccountRow(
-        domain: Account(
-          id: accountId, name: "stub", type: .bank, instrument: .AUD)
-      )
-      .insert(database)
-      try CategoryRow(domain: category).insert(database)
-      try EarmarkRow(
-        domain: Earmark(id: earmarkId, name: "Trip", instrument: .AUD)
-      )
-      .insert(database)
-      try EarmarkBudgetItemRow(
-        domain: EarmarkBudgetItem(
-          id: budgetItemId, categoryId: categoryId,
-          amount: InstrumentAmount(quantity: 100, instrument: .AUD)),
-        earmarkId: earmarkId
-      ).insert(database)
-      try TransactionRow(
-        id: txnId,
-        recordName: TransactionRow.recordName(for: txnId),
-        date: Date(),
-        payee: nil,
-        notes: nil,
-        recurPeriod: nil,
-        recurEvery: nil,
-        importOriginRawDescription: nil,
-        importOriginBankReference: nil,
-        importOriginRawAmount: nil,
-        importOriginRawBalance: nil,
-        importOriginImportedAt: nil,
-        importOriginImportSessionId: nil,
-        importOriginSourceFilename: nil,
-        importOriginParserIdentifier: nil,
-        encodedSystemFields: nil
-      ).insert(database)
-      try TransactionLegRow(
-        id: legId,
-        recordName: TransactionLegRow.recordName(for: legId),
-        transactionId: txnId,
-        accountId: accountId,
-        instrumentId: Instrument.AUD.id,
-        quantity: -1000,
-        type: TransactionType.expense.rawValue,
-        categoryId: categoryId,
-        earmarkId: nil,
-        sortOrder: 0,
-        encodedSystemFields: nil
-      ).insert(database)
-      // Trigger that aborts the budget-item DELETE that the
-      // category-delete path runs after the leg UPDATE.
+      try Self.insertCategoryDeleteParents(database: database, ids: ids, category: category)
+      try Self.insertCategoryDeleteTransaction(database: database, ids: ids)
       try database.execute(
         sql: """
           CREATE TRIGGER fail_category_delete_budget
@@ -216,9 +178,75 @@ struct CoreFinancialGraphRollbackTests {
           END;
           """)
     }
+  }
+
+  nonisolated private static func insertCategoryDeleteParents(
+    database: Database,
+    ids: CategoryDeleteFixtureIds,
+    category: Moolah.Category
+  ) throws {
+    try AccountRow(
+      domain: Account(
+        id: ids.accountId, name: "stub", type: .bank, instrument: .AUD)
+    ).insert(database)
+    try CategoryRow(domain: category).insert(database)
+    try EarmarkRow(
+      domain: Earmark(id: ids.earmarkId, name: "Trip", instrument: .AUD)
+    ).insert(database)
+    try EarmarkBudgetItemRow(
+      domain: EarmarkBudgetItem(
+        id: ids.budgetItemId, categoryId: ids.categoryId,
+        amount: InstrumentAmount(quantity: 100, instrument: .AUD)),
+      earmarkId: ids.earmarkId
+    ).insert(database)
+  }
+
+  nonisolated private static func insertCategoryDeleteTransaction(
+    database: Database,
+    ids: CategoryDeleteFixtureIds
+  ) throws {
+    try TransactionRow(
+      id: ids.txnId,
+      recordName: TransactionRow.recordName(for: ids.txnId),
+      date: Date(),
+      payee: nil,
+      notes: nil,
+      recurPeriod: nil,
+      recurEvery: nil,
+      importOriginRawDescription: nil,
+      importOriginBankReference: nil,
+      importOriginRawAmount: nil,
+      importOriginRawBalance: nil,
+      importOriginImportedAt: nil,
+      importOriginImportSessionId: nil,
+      importOriginSourceFilename: nil,
+      importOriginParserIdentifier: nil,
+      encodedSystemFields: nil
+    ).insert(database)
+    try TransactionLegRow(
+      id: ids.legId,
+      recordName: TransactionLegRow.recordName(for: ids.legId),
+      transactionId: ids.txnId,
+      accountId: ids.accountId,
+      instrumentId: Instrument.AUD.id,
+      quantity: -1000,
+      type: TransactionType.expense.rawValue,
+      categoryId: ids.categoryId,
+      earmarkId: nil,
+      sortOrder: 0,
+      encodedSystemFields: nil
+    ).insert(database)
+  }
+
+  @Test
+  func categoryDeleteRollsBackOnFailure() async throws {
+    let database = try ProfileDatabase.openInMemory()
+    let categoryRepo = GRDBCategoryRepository(database: database)
+    let ids = CategoryDeleteFixtureIds()
+    try await seedCategoryDeleteFixture(in: database, ids: ids)
 
     do {
-      try await categoryRepo.delete(id: categoryId, withReplacement: nil)
+      try await categoryRepo.delete(id: ids.categoryId, withReplacement: nil)
       Issue.record("delete should have thrown but did not")
     } catch {
       // Expected.
@@ -227,16 +255,16 @@ struct CoreFinancialGraphRollbackTests {
     // Category survives, leg keeps its category_id, budget item still
     // references the category — the entire transaction rolled back.
     try await database.read { database in
-      let surviving = try CategoryRow.filter(CategoryRow.Columns.id == categoryId)
+      let surviving = try CategoryRow.filter(CategoryRow.Columns.id == ids.categoryId)
         .fetchOne(database)
       #expect(surviving != nil)
       let leg = try #require(
-        try TransactionLegRow.filter(TransactionLegRow.Columns.id == legId)
+        try TransactionLegRow.filter(TransactionLegRow.Columns.id == ids.legId)
           .fetchOne(database))
-      #expect(leg.categoryId == categoryId)
+      #expect(leg.categoryId == ids.categoryId)
       let budgetItem =
         try EarmarkBudgetItemRow
-        .filter(EarmarkBudgetItemRow.Columns.id == budgetItemId)
+        .filter(EarmarkBudgetItemRow.Columns.id == ids.budgetItemId)
         .fetchOne(database)
       #expect(budgetItem != nil)
     }
